# apps/api/internal/platform/postgres/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package db

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type Alert struct {
	ID            pgtype.UUID        `json:"id"`
	RuleID        pgtype.UUID        `json:"rule_id"`
	TransactionID pgtype.UUID        `json:"transaction_id"`
	Status        string             `json:"status"`
	Severity      string             `json:"severity"`
	CreatedAt     pgtype.Timestamptz `json:"created_at"`
	UpdatedAt     pgtype.Timestamptz `json:"updated_at"`
}

type OutboxEvent struct {
	ID          int64              `json:"id"`
	Aggregate   string             `json:"aggregate"`
	EventType   string             `json:"event_type"`
	Payload     []byte             `json:"payload"`
	OccurredAt  pgtype.Timestamptz `json:"occurred_at"`
	PublishedAt pgtype.Timestamptz `json:"published_at"`
}

type Rule struct {
	ID        pgtype.UUID        `json:"id"`
	Name      string             `json:"name"`
	Field     string             `json:"field"`
	Operator  string             `json:"operator"`
	Value     string             `json:"value"`
	Severity  string             `json:"severity"`
	Enabled   bool               `json:"enabled"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
}

type Transaction struct {
	ID          pgtype.UUID        `json:"id"`
	AccountID   string             `json:"account_id"`
	AmountMinor int64              `json:"amount_minor"`
	Currency    string             `json:"currency"`
	Direction   string             `json:"direction"`
	Status      string             `json:"status"`
	OccurredAt  pgtype.Timestamptz `json:"occurred_at"`
	CreatedAt   pgtype.Timestamptz `json:"created_at"`
}