# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: FinWatch API
  version: 0.1.0
  description: |
    REST contract for FinWatch — an open-source reference implementation for
    near-real-time transaction monitoring and operational alert workflows using
    synthetic data.

    This is the contract-first source of truth: handlers must conform to it.
    Only operational endpoints exist in the bootstrap; feature endpoints
    (transactions, alerts, rules) are added in their own issues.

    Conventions:
      * Monetary values are integer minor units (never floating point).
      * Timestamps are RFC 3339 (UTC) at the API boundary.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
    description: Local development

tags:
  - name: health
    description: Operational liveness and readiness probes.
  - name: transactions
    description: Synthetic transaction records.
  - name: alerts
    description: Operational alerts and their lifecycle.

paths:
  /health/live:
    get:
      tags: [health]
      operationId: getLiveness
      summary: Liveness probe
      description: Returns 200 while the process is running. No dependencies are checked.
      responses:
        "200":
          description: Process is alive.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LivenessStatus"
              example:
                status: ok
  /health/ready:
    get:
      tags: [health]
      operationId: getReadiness
      summary: Readiness probe
      description: Returns 200 only when all downstream dependencies are reachable.
      responses:
        "200":
          description: Service is ready to accept traffic.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadinessStatus"
              example:
                status: ready
        "503":
          description: A dependency is unavailable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: NOT_READY
                  message: database unreachable
  /transactions:
    get:
      tags: [transactions]
      operationId: listTransactions
      summary: List synthetic transactions
      description: |
        Returns a page of synthetic transactions, most-recent first, using
        opaque keyset (cursor) pagination. All data is synthetic.
      parameters:
        - name: limit
          in: query
          required: false
          description: Maximum items to return (1-200, default 50).
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: cursor
          in: query
          required: false
          description: Opaque cursor from a previous response's nextCursor.
          schema:
            type: string
      responses:
        "200":
          description: A page of transactions.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionPage"
        "400":
          description: Invalid query parameters.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: INVALID_QUERY
                  message: limit must be between 1 and 200
  /alerts:
    get:
      tags: [alerts]
      operationId: listAlerts
      summary: List alerts
      description: |
        Returns a page of alerts, most-recent first, with an optional status
        filter and opaque keyset (cursor) pagination.
      parameters:
        - name: status
          in: query
          required: false
          description: Filter by lifecycle status.
          schema:
            type: string
            enum: [open, acknowledged, resolved]
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: cursor
          in: query
          required: false
          description: Opaque cursor from a previous response's nextCursor.
          schema:
            type: string
      responses:
        "200":
          description: A page of alerts.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertPage"
        "400":
          description: Invalid query parameters.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /alerts/{id}:
    get:
      tags: [alerts]
      operationId: getAlert
      summary: Get an alert by id
      parameters:
        - $ref: "#/components/parameters/AlertId"
      responses:
        "200":
          description: The alert.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
        "404":
          description: Alert not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /alerts/{id}/acknowledge:
    post:
      tags: [alerts]
      operationId: acknowledgeAlert
      summary: Acknowledge an alert (open → acknowledged)
      description: Idempotent — returns the alert unchanged if already acknowledged.
      parameters:
        - $ref: "#/components/parameters/AlertId"
      responses:
        "200":
          description: The updated alert.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
        "404":
          description: Alert not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: The alert is not in a state that allows acknowledgement.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: INVALID_TRANSITION
                  message: alert is not in a state that allows this transition
  /alerts/{id}/resolve:
    post:
      tags: [alerts]
      operationId: resolveAlert
      summary: Resolve an alert (acknowledged → resolved)
      description: Idempotent — returns the alert unchanged if already resolved.
      parameters:
        - $ref: "#/components/parameters/AlertId"
      responses:
        "200":
          description: The updated alert.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
        "404":
          description: Alert not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: The alert is not in a state that allows resolution.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  schemas:
    LivenessStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [ok]
    ReadinessStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [ready]
    Transaction:
      type: object
      description: A single synthetic transaction. Amounts are integer minor units.
      required: [id, accountId, amountMinor, currency, direction, status, occurredAt]
      properties:
        id:
          type: string
          format: uuid
        accountId:
          type: string
          description: Synthetic account identifier.
        amountMinor:
          type: integer
          format: int64
          description: Amount in integer minor units (e.g. cents). Always positive.
        currency:
          type: string
          description: ISO 4217 currency code.
          pattern: "^[A-Z]{3}$"
        direction:
          type: string
          enum: [debit, credit]
        status:
          type: string
          enum: [posted]
        occurredAt:
          type: string
          format: date-time
          description: RFC 3339 (UTC) time the transaction occurred.
    TransactionPage:
      type: object
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        nextCursor:
          type: [string, "null"]
          description: Opaque cursor for the next page, or null when there are no more.
    Alert:
      type: object
      required: [id, ruleId, transactionId, status, severity, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        ruleId:
          type: string
          format: uuid
        transactionId:
          type: string
          format: uuid
        status:
          type: string
          enum: [open, acknowledged, resolved]
        severity:
          type: string
          enum: [low, medium, high, critical]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AlertPage:
      type: object
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Alert"
        nextCursor:
          type: [string, "null"]
          description: Opaque cursor for the next page, or null when there are no more.
    Error:
      type: object
      description: Canonical error envelope returned by all error responses.
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable, machine-readable error code (e.g. NOT_FOUND).
            message:
              type: string
              description: Human-readable description; safe to display.
  parameters:
    AlertId:
      name: id
      in: path
      required: true
      description: Alert identifier.
      schema:
        type: string
        format: uuid
  responses:
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: NOT_FOUND
              message: resource not found